The entirety of the project so far
CLASSIC, this is a smashed out mini showcase of some of the realm moving, divine influences, item collecting, and movement mechanics in wanderlustRPG but just the tiny little mechanics showcases, so where this project has little lines and nodes the main game has full levels (the nodes), the bosses are little icons that smack around instead of actual bosses ect. So as I add more to this it'll basically show the pacing and mapping of the game without any assets or game stuff. Pretty cool as a practice project though, lets me get some of these things "prototyped" without needing real.. anything.